Change the way cutoff for rewards is calculated (#72)
### Description

In this version, we had added a `nextDistribution` field to Firebase. For some reason, it was under the `content` item, which made no sense and caused the app to crash in older versions because it was expecting a string before.

This `nextDistribution` was the date when the next CELO rewards distribution would be made. This is always the following Monday, since that's the cutoff date, so instead of having that value on Firebase now we are just calculating it in the client directly. This will make sure it's always up to date and we don't forget to update it as well.

### Tested

Manually.

### Related issues

- Fixes #64

### Backwards compatibility

Yes