Wait for both navigator refs to be initialized (#129)
### Description

Sometimes the deep link gets missed when it gets sent while the app is initializing. To reproduce this swipe the app away and send a deep link to it.

### Other changes

I also increased the number of retries we wait for the navigator to be initialized just to be safe.

### Tested

Swipe the app away and launch it using a deep link using the script ./scripts/adb-deep-link.sh which sends a pay inte

### Related issues

- Fixes #128 

### Backwards compatibility

Yes.