Add flow for recovering access to account if Redux store is wiped (#195)
When we detect that there is no account stored on the Redux store but there is one on the geth wallet, it means that the Redux store was unexpectedly wiped. To recover from this situation, this PR adds a new screen notifying the user that there's been a user and letting them know that they'll need to input their name and phone number again.

We'll be able to skip the name step in the future when CIP-8 lands.

https://user-images.githubusercontent.com/6062888/113381557-a6826a00-9355-11eb-8291-3d983c9b5b6f.mov

- Added an option to the hidden dev settings to erase the Redux store in order to be able to test this.

Manually using the hidden dev setting.

- Fixes #189

N/A