Force y18n version 5.0.5 (#234)
### Description

A new vulnerability was published recently in the y18n package: https://www.npmjs.com/advisories/1654

I tried in PR https://github.com/celo-org/wallet/pull/177 to upgrade the packages that are using the old version of y18n but it quickly went out of control to the point where the codegen library would break other parts of the code. I've spent some time trying to upgrade everything except for the codegen lib (which is needed in the notification service) but the change became really messy. So, I decided to just use the force resolution mechanism for now.

### Other changes

None

### Tested

Build and run tests locally.

### Related issues

- Fixes #174 

### Backwards compatibility

Yes.