Prevent repeated code from being pasted in Android (#391)
### Description

When codes are autopasted on Android, they were sometimes being duplicated. I fixed this in two places:
- In the SMS receiver listener.
- In the reducer so as to prevent other duplication issues in the future. This code really needs a big refactoring to make it more consistent.

### Other changes

N/A

### Tested

Locally. It's hard to reproduce consistently so I changed the code temporarily to test so that the paste code callback is called with the same code 3 times and made sure that only the first one is handled.

### How others should test

Verify using Android and make sure that the autopasted codes are not repeated.

### Related issues

- Fixes #390

### Backwards compatibility

N/A