Copy the start of the indexer code into a separate package.
This is copy-paste from @gnardini's indexer branch at
https://github.com/celo-org/wallet/tree/indexer. The indexing code in
that branch is built in the notification-service package. We plan to
eventually refactor the notification fuctionality out of that package,
but until we do it's difficult to collaborate on the indexer code.

To facilitate immediate collaboration, we're copy-pasting the inter-mixed
indexer+notification-service code into a new package. A separate follow-on
PR will remove the notification functionality from the indexer package.