Show a different text in the feed for the new savings rewards (#656)
### Description

Updated the feed for the new rewards program. The title text is now different (`Earned CELO` -> `CELO Network`).

I created a new array in Firebase called `rewardsSenders`.
In the past for CELO rewards we added new entries to the `addressesExtraInfo` node with a boolean indicating that it's a reward, but that one is growing a lot with a lot of identical elements which have the same name & image (the name is being ignored and set client-side). By having this array instead we can leave the `addressesExtraInfo` node for the initial purpose (special addresses like Moola or cash in providers) and not dirty it with more conditionals.

![simulator_screenshot_60C61F6A-AF13-4756-959E-D8C25D978997](https://user-images.githubusercontent.com/6062888/124066554-42804800-da0f-11eb-95b2-b9ae0bed274a.png)

### Other changes

Converted the recipients reducer to redux toolkit.

### Tested

Set one of my test addresses as a rewards sender to test.

### How others should test

Hard to test this one. 
You can send me your address and I'll make a transaction to you from a fake rewards sender address. Once you have an address. I can also set any address you want as a 'rewards sender address'.
For any address marked as a rewards address, all incoming transactions from it should say `CELO Network` in the title and `Weekly Earnings` in the subtitle.
In the transaction detail for these rewards, the `Learn more` button should redirect to the rewards screen.

### Related issues

- Fixes #655

### Backwards compatibility

N/A