Hide keyboard when selecting currency to exchange (#856)
### Description

Fix for this issue found by QA: https://github.com/celo-org/wallet/issues/851

In short, in the exchange amount screen, when a user has the keyboard open to input an amount and they tap on the header the balance picker was showing below the keyboard. We're now hiding the keyboard when the header is tapped.

### Other changes

N/A

### Tested

On iOS and Android

### How others should test

Go to the exchange screen and try to change the currency to do the exchange with. The keyboard should close when tapping on the header.

### Related issues

- Fixes #851

### Backwards compatibility

N/A