CTA Localization / Text Accessibility (#855)
### Description

Allows text of various lengths and sizes to display appropriately in the notification box / call to action display.

### Other changes

Test Snapshots updated

| | iOS Before | iOS After | Android Before | Android After |
| ---- | ---- | ---- | ---- | ---- |
| CTA | [Before](https://drive.google.com/file/d/1xwjGwe2j_xC5Tpdbt3Hr8MbHot6WVkcx/view?usp=sharing) | [After](https://drive.google.com/file/d/1jUEDrJdt8wI4EBtE_A0WVUmj-vDykiAx/view?usp=sharing) | [Before](https://drive.google.com/file/d/15wWWWyQud1BVaEfhik24lcgIkv60lWkB/view?usp=sharing) | [After](https://drive.google.com/file/d/1jbOM7GRyGEzfBvJwCPDhpmweRze4bWHZ/view?usp=sharing) |

### Tested

Tested locally on iOS and android devices.

### How others should test

- Launch the app with default text sizing and insure the CTA displays correctly.
- Chance the text size and display size (android only) relaunch the app and insure that the CTA displays correctly.

### Related issues

- Fixes #149 #579

### Backwards compatibility

As this impacts styling only it is backwards compatible.

### Known Issues

Currently device width and height is only calculated on app load from [App.tsx](https://github.com/celo-org/wallet/blob/main/packages/mobile/src/app/App.tsx). This means that launching the app, changing system settings and bring the app to the foreground does not recalculate these variables. One solution for this would be move to use [https://reactnative.dev/docs/usewindowdimensions](https://reactnative.dev/docs/usewindowdimensions) and setup an [event listener for window resizing](https://stackoverflow.com/questions/19014250/rerender-view-on-browser-resize-with-react); however, that is outside of the scope of this PR.  Currently the workaround for resizing is to close and relaunch the app.