Fix `postinstall` script (#1133)
### Description
The `postinstall` script needed to be updated. The package that was part of the monorepo, which previously would prompt the unzipping of the `Geth` file, is no longer part of our repo. Hence we must manually unzip this to avoid build errors.

### Other changes
N/A

### Tested
Manually

### How others should test
N/A

### Related issues
N/A

### Backwards compatibility
Yes