Fix Flaky Tests (#1175)
### Description

Prior to this PR these tests were asserting based on the balance before and after the exchange. This was causing issues when multiple tests runs were active in the pipeline. With this PR exchange tests are use the time of the exchange and the amount to assert that they were completed successfully. 

### Other changes

This also increases the time out for the 'Then Should Display Providers' for `OnRamps.spec.js` which was occasionally failing in CI due to slower rendering in CI.

### Tested

Observed CI runs and run locally.

### How others should test

N/A