Remove Clear Input Button From Phone Number Input (#1283)
### Description

Currently the clear input button does not work on the phone number input. I tried to get it to work, but didn't find an easy solution. As a temporary solution I've disabled it from showing to prevent the inability to clear from surfacing to users.

### Tested

Tested locally

### How others should test

Tested by typing into the phone number input and observing that the clear button is not rendered.

### Related issues

- "Fixes" #461 #576

### Backwards compatibility

This change is backwards compatible.