Duplicate Send Amount flow screens (#1301)
### Description

First PR for the send amount flow. There are a lot of lines "added" but almost no new real code.

This one is not changing any logic, only copying some of the classes which will have the most changes so that we have two flows: A legacy one and the new one. For now, the code will only use the legacy flow but once we finish the send flow implementation we will change it to use the updated flow.

This has the advantage that we aren't merging one huge PR at any point and it's easier to implement things more incrementally.

### Other changes

N/A

### Tested

Ran the app, no visible changes should happen.


### How others should test

N/A

### Related issues

- Part of #1126

### Backwards compatibility

N/A