Fetch token balances using Blockscout (#1386)
### Description

Before this PR when fetching balances we were querying the balance directly from the token contracts.
This has the advantage that it's querying directly from Forno or other nodes that are more likely to be up to date than blockscout, which sometimes has a delay to update balances.

However, the disadvantage was that we were making one request per each token. If there are not a lot of tokens this is fine, but there are 42 tokens already on https://github.com/Ubeswap/default-token-list/blob/master/ubeswap.token-list.json and this number will continue to grow. Making so many requests each time we want to fetch balances looks like too much.

Another possible solution longer-term is indexing and updating the balances of our users ourselves using the indexer. This would give us control over the latency for block updates while not having to make so many requests at the same time.

### Other changes

- Created an action which re-fetches the token balances and called it in the places where the old ones used to be called.
- Created an error action to be used in the UI when there are errors in future PRs.

### Tested

Manually and with unit tests.

### How others should test

Nothing to test directly here, just make sure the balances show up and are up to date after making a transfer.
### Related issues

- Part of #1126

### Backwards compatibility

N/A