remove direct calls to i18n.t when useTranslation hook is available (#1563)
### Description

We should avoid calling the `t` function from i18n directly, because the imported `i18n` instance could be uninitialised depending on when it was imported. This causes empty or raw translation keys to be rendered in edge cases (see https://github.com/valora-inc/wallet/issues/1557 and https://github.com/valora-inc/wallet/pull/1561)

This PR makes the easy changes, to use the `t` function from the useTranslation hook when possible. We will need to find some ways of handling places where we do translations in .ts files and static functions (like navigation options) in the future.

### Other changes

N/A

### Tested

Manually tested the affected screens to see translations are still coming through

### How others should test

Same as above

### Related issues

N/A

### Backwards compatibility

N/A