Bump app version to 1.24.0 (#1678)
### Description

Discussing with Jean while documenting the release process, we thought it was a good idea to apply the version bump to main _after_ every release, rather than bumping the version immediately before every release.

This means that after a release branch is cut, the app version in main is bumped and new features merged after that are part of the next release version.

### Other changes

n/a

### Tested

not needed