Fix fee calculation for non cUSD tokens (#1733)
### Description

There were a couple of issues.
First, the token to use as fee currency was being chosen at random. Not a huge deal I guess, but switched back to using the token with highest balance for consistency.
The more important issue is that we were sending to the `FeeDrawer` the `totalFee` in USD but also sending the fee currency and the exchange rate, so in the end we were transforming the value and ended up showing something inaccurate. Since we're already calculating the fee in USD, we are not just passing the currency as `Currency.Dollar` and letting the `FeeDrawer` component transform it into the local currency normally (`CurrencyDisplay` ends up doing that).

### Other changes

N/A

### Tested

I did a TDD approach here, if you run the new test with the old code it fails, but after the changes it works.

### How others should test

Try to send using different fee currencies (decided by the highest balance, so try while having more cUSD than cEUR/CELO and then by having more cEUR than cUSD/CELO) and see that the fee currency has an expected value. Not sure how to decide what "expected value" is though 🤔 

### Related issues

- Fixes #1727

### Backwards compatibility

N/A