use https:// instead of git:// security protocol no longer supported (#1744)
### Description

Background: https://github.blog/2021-09-01-improving-git-protocol-security-github/ - the final brownout for this change was today.

The git:// security protocol is no longer supported and our CI was unable to finish fetching dependencies. Update to use https://

### Other changes

N/A

### Tested

No testing is needed, as we are just updating the protocol for fetch one dependency. The app should be unaffected.


### How others should test

As above.

### Related issues

N/A

### Backwards compatibility

Yes