Dynamic total steps number on registration flow screen headers (1759)
### Description

In the onboarding registration flow, the screen headers are hard coded to say something like `step {{step}} of 3`. For the biometrics experiment, we need to allow for an extra registration step so we need to make this dynamic.

Rather than going through each screen and adding an extra condition, i've opted to extract this simple logic in a reusable hook. Unfortunately 2 of our onboarding screens are class components so I couldn't use them there.

### Other changes

In ImportWallet and VerificationEducationScreen, as they are functional componets, move navigation options into `useLayoutEffect` so that `t` function from the `useTranslation` hook can be used. This is consistent with the pattern used in NameAndPicture.

### Tested

Manually

### How others should test

The onboarding flow (create account/restore account) should work as expected and the step numbers should work as expected.

<img width="414" alt="Screenshot 2022-01-12 at 11 47 23" src="https://user-images.githubusercontent.com/20150449/149126263-5ff1010b-eaf2-47e8-acbd-314728703b53.png">


### Related issues

- Relates to #1683 

### Backwards compatibility

Yes