Allow sending transfers with comments in new flow for core tokens (#1769)
### Description

Sending with comment was broken because we were comparing uppercase letters with lowercase ones. Sending with comment is currently not working on the 1.24 release. This fixes it while also adding cREAL as possible token to send with comment.

### Other changes

N/A

### Tested

Manually sending core tokens and not core tokens.

### How others should test

- Send cUSD/cEUR/cREAL and see that you can send a comment with it.
- Send CELO or any other token and check that you can't send a comment with it.

### Related issues

- Fixes #1767
- 
### Backwards compatibility

N/A