Add RU/FR/IT to supported languages (#1797)
### Description

Add RU/FR/IT to supported languages. Please note that the translations for these 3 new languages are currently incomplete (many keys are still using english in the translation files) but now RU/FR/IT should appear as options in the language selection pages on app start/settings screen. 

### Other changes

N/A

### Tested

Manually verify that the new languages can be selected from and the app changes languages on selection.


### How others should test

Same as above

### Related issues

- Fixes #1721 

### Backwards compatibility

Yes