Dapp explorer (#1784)
### Description

DApps Screen.
Designs: https://www.figma.com/file/zdK8XvxNXvTjxnM5V33pHw/dAPP-Directory?node-id=831%3A19659

### Tested

Manually in the Android and iOS emulator

### How others should test

Turn on the feature flag `dappsExplorerEnabled`, then you should be able to interact with the new screen.

### Known Issue
In the new Dapps screen, if the user presses on Bidali (which it's not visible yet) and then goes back, it doesn't return to the Dapps screen.

### Related issues

- Fixes #1506

Screenshots (this is a older version, some wording has changed):
<img width="493" alt="Screen Shot 2022-01-14 at 13 30 09" src="https://user-images.githubusercontent.com/12504748/149550602-a3f16d3f-be3c-4fc7-bc16-00fe30578dac.png">
<img width="502" alt="Screen Shot 2022-01-14 at 13 30 35" src="https://user-images.githubusercontent.com/12504748/149550680-0d944c4d-d4f6-41b1-891c-cd69861466a6.png">
<img width="497" alt="Screen Shot 2022-01-14 at 13 31 58" src="https://user-images.githubusercontent.com/12504748/149550846-17745e75-4996-4b90-95de-0c7dbc73e0ad.png">
<img width="496" alt="Screen Shot 2022-01-14 at 13 32 16" src="https://user-images.githubusercontent.com/12504748/149550885-a65eecd1-fada-4225-9bdb-bd40f4d0826d.png">