Replace user address in dapp URLs (#1822)
### Description

This will allow some dApps to open with the user already 'connected'. The motivating example is CeloTracker, where you can put the address in the URL like `https://celotracker.com/?address={{address}}` and it will show that address's balance.

### Other changes

N/A

### Tested

Manually


### How others should test

No way for QA to test this at this time, but whenever Celo Tracker shows up in the Dapps screen then it should open with the user's account already set when opened

### Related issues

- Fixes #1821

### Backwards compatibility

N/A