Resolve node-fetch to version 2.6.7 which contains security patch to fix CI (#1833)
### Description

Our vulnerabilities CI has identified node-fetch as a new vulnerability and recommends to upgrade to version 2.6.7. This is the error identified https://github.com/advisories/GHSA-r683-j2x4-v87g

Example of the error in CI: https://github.com/valora-inc/wallet/runs/4918866732?check_suite_focus=true

Version 2.6.7 of node-fetch is a security patch update: https://github.com/node-fetch/node-fetch/releases/tag/v2.6.7

Most of our dependencies are using version 2 of node-fetch already so this resolution is low risk. Note that only `graphql-codegen` depended on node-fetch version 1.x.x, it is only used for the `build:gen-graphql-types` script but as far as I know this isn't being actively used.

We need to resolve this to make our CI green again 🟢

### Other changes

N/A

### Tested

N/A


### How others should test

Does not need to be tested

### Related issues

N/A

### Backwards compatibility

Yes