Don't allow sending tokens if exchange rate wasn't loaded successfully (#1871)
### Description

If the exchange rate isn't loaded, then the Send flow can't show the correct amounts. We are now preventing navigation to that screen if the exchange rate wasn't loaded.

### Other changes

N/A

### Tested

Manually and with unit tests.

### How others should test

To test this change you need the exchange rate query to fail. I did this by editing the code directly, but I don't know how others can reproduce this really :(

### Related issues

- Fixes #1870

### Backwards compatibility

N/A