Allow reclaiming invites sent with any token (#1860)
### Description

Currently only cUSD and maybe cEUR invites can be reclaimed.
With these changes, invites sent with any token will show up and be available for reclaiming.

Note that most of the LoC change were snapshot tests. Most of the files in the reclaim flow had only snapshot tests.
I updated the tests for the main screen `ReclaimPaymentConfirmationScreen` but just updated snapshots in the rest.

<img width="374" alt="Screen Shot 2022-01-26 at 14 54 50" src="https://user-images.githubusercontent.com/6062888/151219250-07b1469d-44c8-4439-8460-f86b503613e6.png">

### Other changes

- Turned a few classes into functional components.

### Tested

Manually and with unit tests.

### How others should test

- Invite someone using cREAL.
- Reclaim the invite.

### Related issues

- Fixes #1859

### Backwards compatibility

N/A