Show featured Dapp in Dapps screen (#1883)
### Description

Added a featured Dapp at the top of the Dapps screen.
This screen expects a `featured` dapp to come in the response which is not currently there. This will be added in #1831

![Simulator Screen Shot - iPhone 11 - 2022-01-28 at 17 36 20](https://user-images.githubusercontent.com/6062888/151617745-7f655bb4-2d46-4fe8-b5ee-4d76bc9b6a63.png)

### Other changes

Added some tests.

### Tested

Manually hard-coding the data of the featured dapp and with the unit tests.

### How others should test

Open the screen. If there is no featured dapp the screen should look the same, otherwise the featured dapp should show up at the top.

### Related issues

- Fixes #1827

### Backwards compatibility

N/A