feat: Resolve Kit Integration (Nom Space integration) (#1872)
### Description

The `RecipientPicker` is now integrated with the [resolve-kit](https://github.com/valora-inc/resolve-kit), which includes the integration with NomSpace.

### Tested

Manually with an iOS emulator

### How others should test

Running in alfajores is a bit hard to test because I don't know any nom space domain there. I will try to find some so it's easier to test.
But running in mainnet, you should be able to test this by typing the nom domain in the send or request screen. 

### Related issues

- Fixes #1804

### Screenshots
![Simulator Screen Shot - iPhone 12 - 2022-01-27 at 17 43 19](https://user-images.githubusercontent.com/12504748/151588225-00054a12-675a-49d2-8907-889a9aae872f.png)
![Simulator Screen Shot - iPhone 12 - 2022-01-27 at 17 43 23](https://user-images.githubusercontent.com/12504748/151588234-95082d9d-943a-4392-bd9e-3a3af80a9357.png)