Fix fee estimation issue when sending through a deep link (#1912)
### Description

If a user opens a payments deeplink they get directed straight to the `SendConfirmation` screen and therefore they didn't have a fee estimation, so the Send button wasn't clickable.

This PR makes it so that the fee estimation is made if it's not already in the Redux store.

### Other changes

N/A

### Tested

Manually

### How others should test

- On a fresh install, open a deep link like `celo://wallet/pay?address=0xB2539a7dcc69a0B28c7Ca73f1D2FF06b9d59E5b9&displayName=Test&currencyCode=KES&amount=0.1`
- See that you can send correctly.

### Related issues

- Fixes #1911

### Backwards compatibility

N/A