Add some delay for biometry authentication to allow native verification animation to run fully (#1922)
### Description

Add delay in biometry get/set functions so that the native animations can run fully before moving the user on. Without the delay, the animation can be running while the app is navigating to the next screen already.

With a 800ms delay, the animation runs like this:

https://user-images.githubusercontent.com/20150449/152314145-9f21302e-1f6f-4123-823b-889b1930b645.MP4


### Other changes

N/A

### Tested

Unit tests and manually


### How others should test

In the onboarding flow when enabling biometrics, and in flows where you need biometry to get the pin (recovery phrase screen, send flow, change pin, etc.) see that the biometric verification animation run to the end before the app moves on.

### Related issues

N/A

### Backwards compatibility

N/A