Make error logging consistent so that errors are grouped correctly in Sentry (#1967)
### Description

Now that we are sending errors to Sentry, consistency is more important than ever so that similar errors can be grouped together to help us find issues/trends faster.

We've noticed some strange errors come through on Sentry like `[object object]` and `undefined`. I think this is caused by what we are sending to the `Sentry.captureException` function. 
- Sometimes we pass a whole error object to the `message` which is expecting a string, which shows up as `[object object]`.
- Sometimes we had ```Logger.error(`${TAG}:fetchSimplexPaymentData`, error.message)``` which would give `undefined` if `error` didn't have a `message` property.

<img width="602" alt="Screenshot 2022-02-08 at 16 40 27" src="https://user-images.githubusercontent.com/20150449/153021848-febe3b5e-646a-42b0-bb42-f912aad944e6.png">


In this PR:
- explicitly turn the message sent to Sentry to a string, as Typescript currently doesn't warn us against passing objects so we can't really prevent this in the future until TS is upgraded
- go through the app and make sure the logger error function is called like `Logger.error(<some tag>, <some message>, <error object?>)`

### Other changes

N/A

### Tested

Manually


### How others should test

This does not need to be tested

### Related issues

N/A

### Backwards compatibility

N/A