Fixes in the verification flow (#2003)
### Description

Some context:
On May 2021 we had a huge issue with verification. There was a huge refactor that brought a bunch of issues that we ended up reverting. _Before_ reverting though we fixed a lot of bugs related to it. In the middle of those bugfixes there were some other bugfixes that were pre-existing! However, when we reverted we reverted everything so we lost those fixes.

Here's the PR from that time: https://github.com/valora-inc/wallet/pull/432

In this PR I'm bringing back two bugfixes that will help with the verification issues.

- The first one is calling a lock we were already using properly. We were using a lock in a section on the code but we were not actually `awaiting` it so it was useless. There was a delay there to "fix" the issue but it didn't really fix it. The lock should now work properly.
- The second one is adding another lock in the code that receives the verification codes. Without it, the UI glitches out a bunch and there may be other issues as well, since all the codes pasted try to get inserted in the same spot. This happens on Android where codes are auto-pasted. With this change, all codes should be pasted in a different position.

This fixes issues for new users, but we also need to run a migration to fix the redux store state for existing users that show up as verified when they are not.

### Other changes

N/A

### Tested

Manually on both OSes. I also modified the code a bit to send the three codes instantly to make sure they are handled correctly.


### How others should test

Try verifying in Android and iOS several times. Try all the edge cases you can think of, including but not limited to:
- Copying the same code several times.
- Pasting the codes very quickly.
- Asking codes to be resent.

### Related issues

- Part of #1906
This doesn't fix the issue completely, it fixes it for new users but we still need to fix the internal state for existing ones.

### Backwards compatibility

N/A