fix: Bump payments-sdk to address gas limit issue (#2218)
### Description

The payments-sdk needed to be updated to prevent a gas limit issue on mainnet

### Other changes

We also commented out the country info for now

### Tested

Was able to send merchant payments with the new sdk.

### How others should test

Running on mainnet go to https://prod-celo-demo-store.demo.firstdag.com/, pick a product, scan the QR code and verify it sends.

### Backwards compatibility

Yes