Use a different fee currency if there's not enough to pay for gas (#2232)
### Description

In the places we (or other dapp developers) don't set a feeCurrency, the app tries to use CELO if the user has enough balance. The problem is that  we check for > 0, not for > gasCost, so if the user has a very small amount of CELO (less than needed to send a tx) then we still try to use CELO and the tx fails.

This has already been affecting some users: https://sentry.io/organizations/valora-inc/issues/3004252806/?project=1250733&query=is%3Aunresolved+gas and the only workaround is to either get rid of their whole CELO balance or getting more CELO.

We are now only using a feeCurrency if the user has enough to pay for gas with it.

### Other changes

N/A

### Tested

With unit tests and manually

### How others should test

- Get an account with a very small amount of CELO (<0.000001)
- Invite someone using the Send flow.
- Try to reclaim the invite. Reclaiming should be successful.

### Related issues

- Fixes #2231

### Backwards compatibility

N/A