Update known vulnerabilities (#2270)
### Description

There is a new vulnerability alert for simple-plist https://github.com/advisories/GHSA-gff7-g5r8-mg8m however the issue is not yet resolved from the project. For now ignore the vulnerability to unblock the CI, but we should bump the resolved version of simple-plist once [issue #60](https://github.com/wollardj/simple-plist/issues/60) is resolved.

### Other changes

N/A

### Tested

N/A


### How others should test

N/A
### Related issues
N/A

### Backwards compatibility

Yes