Update translation.json (#2269)
Changes made:

"superchargingDescription": "You’re getting {{apy}}% annually. Claim your rewards here on Tuesday.”

->

"You’re getting {{apy}}% annually. You’ll get a weekly notification when your rewards are ready, come back here to claim them.",

Is it possible to do a line break between the two sentences? If so that would be ideal.

### Description

_A few sentences describing the overall effects and goals of the pull request's commits.
What is the current behavior, and what is the updated/expected behavior with this PR?_

### Other changes

_Describe any minor or "drive-by" changes here._

### Tested

_An explanation of how the changes were tested or an explanation as to why they don't need to be._


### How others should test

_Does this need to be tested by QA in the next release cycle? If so please give a brief explanation of how to test these changes._

### Related issues

- Fixes #[issue number here]

### Backwards compatibility

_Brief explanation of why these changes are/are not backwards compatible._