Display invitee from contact list in home feed, rather than "unknown" (#2265)
### Description

Currently the transaction item for an invite sent displays as from "Unknown" because the transaction is going to the escrow contract address instead of a known contact address.

In this PR:
- add redux variable to track invitation transaction details. due to the transaction feed being updated regularly, it is nice to keep this persisted redux variable separately so that we don't need to refresh the invite details
- if the transaction feed item is a sent invite, get the recipient info from the escrow contract
- when displaying the transaction feed item, use the fetched recipient details from the escrow contract

<img width="437" alt="Screenshot 2022-03-29 at 16 24 51" src="https://user-images.githubusercontent.com/20150449/160633963-a8206771-32e6-47b4-865b-d3afb36775cb.png">

Note: we still rely on the contacts cache to display the invitee name, this cache is only refreshed when the user accesses their contacts (e.g. when on the send screen) - this means the invitee will still display as "Unknown" if the user has reinstalled the app/has not accessed their contacts list yet.

### Other changes

N/A

### Tested

Manually, unit tests to come

### How others should test

Send an invitation, then observe the transaction item on the home feed should display the correct invitee name for the transaction

### Related issues

- Fixes #2215 

### Backwards compatibility

Yes