Ensure automerge Crowdin translations automation does not modify base translations (#2285)
### Description

The Crowdin translation PR should not modify the base translations. Sometimes when the Crowdin branch is open for a long time and there are new PRs merged that modify the base translations, weird things happen and the base translations are wrongly updated. e.g. https://github.com/valora-inc/wallet/pull/2276

In this PR, check that only the translated languages are modified before automerging. Otherwise, leave a "changes requested" review so that it is easier to see that manual intervention is needed.

### Other changes

N/A

### Tested

Manually


### How others should test

N/A

### Related issues

N/A

### Backwards compatibility

Yes