E2E tests for multi-token send flow (#2311)
### Description

Updates the E2E tests to use the multi-token send flow, built on top of #2289.

### Other changes

Additional tests for deeplink payments.

### Tested

Tested locally on iOS and in CI on Android

### How others should test

This does not need to be tested by QA. For reviewers please check that the e2e tests are passing in CI and that the tests run locally on Android and iOS.

### Related issues

- Fixes #1775

### Backwards compatibility

Yes