Fix home cards (#2354)
### Description

Supercharge notifications are handled in-app instead of fetching them from admin-dashboard.
There are 3 use cases:
* There are rewards to be claimed (see screenshot 1)
* User doesn't meet all the conditions to supercharge (no enough funds or phone number not verified) (see screenshot 2)
* User has all the conditions to supercharge (see screenshot 3)

In every case, if the user clicks on the action, we are taking him to the supercharge page (ConsumerIncentiveHomeScreen).

### Tested

Manually in an emulator, running mock distributions. Also with unit tests.

### How others should test

Just open the app and see the home card. In case you have dismissed the message you would need to reinstall the app in order to see the notification again.

### Related issues

- Fixes #9473

### Backwards compatibility

Compatible, but the notifications may appear twice if we don't do the thing under `Important` (see below)

### Important
We should stop sending the supercharge notifications from the admin-dashboard once this PR is release. To do so, go to https://valora-admin-dashboard.web.app/, and then set as `Max Version` the version of the previous release for the notifications `Supercharge` and `SuperchargeEUR`.

### Screenshot
![output-onlinepngtools](https://user-images.githubusercontent.com/12504748/165310689-6f9ae334-dfed-4543-adaf-8d30220b329c.png)
![output-onlinepngtools (1)](https://user-images.githubusercontent.com/12504748/165310718-b3213eed-664f-460c-8c5a-06ef4b197ea4.png)
![output-onlinepngtools (2)](https://user-images.githubusercontent.com/12504748/165310730-108c4631-493b-43bb-a4b8-cfb2e02cc929.png)