Call setRegistrationProperties endpoint for registrations instead of using Firebase directly (#2351)
### Description

We need to add the country code to the registration node in firebase, to enable region based invite rewards. To decrease our reliance on firebase, we have a new setRegistrationProperties endpoint created that we should call instead.

This PR:
- adds country to the firebase rtdb schema
- replaces the direct calls to Firebase with setRegistrationProperties endpoint for updating registration details
- adds keychain item (without access control, so that no pincode/biometrics are needed to access) for signed message, which is used for authentication for the setRegistrationProperties endpoint.

Note that no changes were made to the RTDB for now, the only extra authentication is on the setRegistrationProperties endpoint.

This signed message is generated when an account is created/imported for new users. For existing users, it will be generated after the first time they enter their pincode for something (to reduce friction) - this means that we will not be updating existing users's registrations node until they have done something to trigger the pincode.

### Other changes

N/A

### Tested

Unit tests and manually

### How others should test

General regression. There should be no change in the user journey.

### Related issues

- Relates to https://github.com/valora-inc/invite-rewards/issues/40

### Backwards compatibility

Yes