bank details screen (#2464)
### Description
This is the screen for user to manually type in their fiat account information. It includes validation and posting to FiatConnect's [Fiat Account endpoint ](https://github.com/fiatconnect/specification/blob/main/fiatconnect-api.md#343-fiat-account-endpoints)

TODO:
1. I made up some of the wordings including error message, placeholder text, system message, plan to get Nitya to review before 
2. We haven't implemented the screen after this one or the one before, the provider selected right now is hardcoded, will need to update the navigation stack once the screens are connected
3. There's a newly added FiatAccountSchema that has not been added to the types yet, need to add to the fiatconnect-types and add that schema

### Tested
unit tests


### How others should test
This component is not accessible in the app yet. It is a future flow that has not been fully built out

### Related issues
- Fixes #2333