Remove clear cache on verification start so that no pincode is needed for phone verification (#2502)
### Description

The password cache is currnetly cleared on start of the verification user journey, with the intent to ensure that the account is unlocked for the whole verification duration. This forces the user to enter their pin, even if they've just entered it for something immediately beforehand (e.g. in the current onboarding flow)

Discussing with Jean, the important thing for the verification flow is that the geth account is unlocked for the maximum time (to ensure all transactions within the rather long verification process would be able to be signed successfully). We already call the geth unlock saga with the `force` flag in this flow, so the unlock period should be at its maximum.

The pincode is only needed for unlocking the geth account, and is not needed during the verification process. Therefore, it should be safe to avoid clearing the password cache here. I tested by shortening the cache time and kicking off the verification process shortly before the cache expires. Can confirm that there was no need to input the pincode again.


### Other changes

N/A

### Tested

Manually


### How others should test

Phone verification in onboarding flow or through setting screen should only prompt for pincode if the user hasn't entered their pincode in the last 5 minutes (and if they haven't force closed the app)

### Related issues

N/A

### Backwards compatibility

Yes