chore: new Keychain based wallet + signer (#2519)
### Description

This adds a new Keychain based wallet + signer to replace the ones based on geth.
The goal is to be able to eventually remove the light client.

The private keys are stored encrypted in the OS Keychain and protected via the user PIN.
This is very similar to how we were keeping the keys secure using the geth keystore.
The main change is the storage: using the keychain instead of the filesystem (via the geth keystore).

I made sure it behaves very closely to the geth based wallet (for instance returning accounts in the order they were created, some parts of the code base rely on this).

Also all signing is now happening on the JS side.
I don't yet know if it impacts the JS performance.

The work to migrate existing user accounts to this new system was done in #2539.
To make it easier to review these changes individually.

### Tested

- Extensive unit tests on the new `KeychainWallet` (and `KeychainSigner` indirectly).
- Manually creating and importing an account.

### How others should test

1. Create a new account or import an existing one.
2. Make sure all things related to signing work as expected
    - sending transaction
    - swapping
    - signing with dapps
    - phone number verification
    - updating the PIN 

### Related issues

- Part of #2392

### Backwards compatibility

Not backward compatible by itself. Backward compatibility added in #2539.