Pincode Set Fix (#2629)
### Description
A follow up to #2593 and #2608. Fixes the incorrect behavior where the wrong error message was served on pins that didn't match. This change moves the `isVerifying` navigation param, which was always passed as false, into the component state.

### Other changes
- Update mobile tests in https://github.com/valora-inc/wallet/commit/5a596f9406f51b31c484dc6e63c3530e88c23539
- Use `Buffer.from` instead of `new Buffer` - [Porting to the Buffer.from()/Buffer.alloc() API](https://nodejs.org/en/docs/guides/buffer-constructor-deprecation/#overview)
- Added e2e test for pincode reentry in https://github.com/valora-inc/wallet/pull/2629/commits/b64a82a9e1347167198b7eddaed8ef695458deda

### Tested
Tested locally on iOS

### How others should test
Perquisites: When creating a pin or changing a pin

1. Enter the first pin.
2. Enter a non-matching second pin.
3. Observe error message should be 'The PINs didn't match'
4. Should be redirected to the first step if second pin doesn't match.

### Related issues
N/A

### Backwards compatibility
Yes