feat: Move recovery phrase into 'settings' (#2758)
### Description

This is an A/B test.

Control: current experience w/ recovery phrase only in the navigation menu

Treatment:
If the Firebase remote config `shouldShowRecoveryPhraseInSettings` is true, the user sees the recovery phrase in settings, not in the navigation menu.
However, if the user hasn't backed up his recovery phrase yet, then the user can see it in the navigation menu as well even though the remote config is true.

<img width="362" alt="Screen Shot 2022-08-04 at 5 33 19 PM" src="https://user-images.githubusercontent.com/47931506/182972579-a63b019f-64ee-4925-82ae-f335cd9bd07b.png"> <img width="362" alt="Screen Shot 2022-08-04 at 5 33 08 PM" src="https://user-images.githubusercontent.com/47931506/182972590-739bd506-ab7c-47db-8ba7-2666418707b4.png"> <img width="362" alt="Screen Shot 2022-08-04 at 5 37 43 PM" src="https://user-images.githubusercontent.com/47931506/182973345-bec3aac3-a06c-43b2-8651-25e51bcac90d.png">

### Other changes
N/A


### Tested
With iOS emulator


### How others should test


### Related issues
- https://linear.app/valora/issue/RET-302/move-recovery-phrase-into-settings


### Backwards compatibility
Yes