feat: skip button for user name during onboarding  (#2890)
### Description

Adds a skip button to the "name" screen of the user onboarding, and
internally leaves the username empty when that is clicked. Clicking skip
gets the user to the next screen.
Written with the intention to allow this feature to be toggled by
Statsig when Statsig is integrated.

### Other changes
None

### Tested

Automated tests are written for `NameAndPicture.tsx`, in which both
showing and not showing skip button are tested.
Manual testing were done for all references to the username with an
empty one, see screenshots below.

### How others should test

After Statsig integration, selecting the control for the skip button
during onboarding username would result in skip button being shown. The
skip button should work and take user to next onboarding screen.

### Related issues

- Fixes
[ACT-436](https://linear.app/valora/issue/ACT-436/skip-button-on-name-screen)

### Backwards compatibility

Compatible; the config values are hard-coded.