ci(e2e): test against multiple iOS versions (#2994)
### Description

I've broken the iOS E2E test into its own reusable workflow, taking the
iOS version as an input - similar to #2968. It is called when a PR is
opened, and run against one iOS version. It's also called against
multiple iOS versions on `main` (when a commit is pushed; when called
from the Github UI; on a schedule, once daily).

### Other changes

I've also changed the name of the JUnit report check (for both iOS and
Android), so we get one report for each OS version. Before, on `main`,
the report would be overridden by subsequent OS version runs.

### Tested

The test suite has been run multiple times against the re-provisioned
MacStadium boxes.

### How others should test

N/A

### Related issues

RET-275

### Backwards compatibility

The required branch protection check will need to be updated from `iOS`
to `iOS (14.5)` before this PR is merged.