fix(home): Fix frozen home screen on iOS (#3048)
### Description



https://user-images.githubusercontent.com/8432644/198121473-82d2dda3-786b-46f8-b012-3a588af89d99.mp4

Fixes
https://valora-app.slack.com/archives/C02N3AR2P2S/p1666741437531239

### Tested

I narrowed down the exact time to use by replacing the entire useAsync
function with a `sleep(<time>)` function and adjusted time until the
issue consistently didn't reproduce. 1000 was the lowest It could go.

### How others should test

On an iOS simulator/device on *main build* attempt to onboard all the
way to the Home screen. Verify that the screen is not frozen AND the
cash in bottom sheet shows up

> you may have to turn on the cash in bottom sheet feature flag to see
it

### Related issues

- Fixes
https://valora-app.slack.com/archives/C02N3AR2P2S/p1666741437531239

Co-authored-by: jh2oman <jacobwaterman@MacBook-Pro.local>