fix: failed CPV during onboarding due to missing signed message (#3064)
### Description

Context:
https://valora-app.slack.com/archives/C044B8ULVC0/p1667413326982599

This PR introduces a loading state to the phone verification start
button while the signed message is not available. Previously on android
this can fail:
- during onboarding the "initialise account" action is dispatched on
this screen load
- the signed message is only generated after the account has been
generated, but this is an asynchronous action
- if the user navigates to the next CPV step before the previous step is
complete, there is no signed message for authentication with CPV, so the
verification will always fail here.


### Other changes
N/A

### Tested

Manually + unit tests. 

### How others should test

On android device:
- create a new account + verify phone number. it should succeed

### Related issues

N/A

### Backwards compatibility

Yes

Co-authored-by: Tom McGuire <Mcgtom10@gmail.com>