chore: add walletconnect v2 to redux (#3161)
### Description

As the title - I kept all the actions and redux state that were removed
in #2324 but had to update the types. I tried to minimise the size of
each PR so this PR only adds redux support (reducer, actions, selectors)
but the actions are not yet used. I've done my best first pass at the
types but they could change after I start hooking up the functionality.

### Other changes

n/a

### Tested

n/a

### How others should test

n/a

### Related issues

- Fixes RET-479

### Backwards compatibility

n/a